+++ src/main/java/info/gratour/common/error/ErrorWithCode.java
@@ -1,10 +1,11 @@
-/** *****************************************************************************
+/**
+ * ****************************************************************************
  * Copyright (c) 2019, 2021 lucendar.com.
  * All rights reserved.
- *
+ * <p>
  * Contributors:
  * KwanKin Yau (alphax@vip.163.com) - initial API and implementation
- * ******************************************************************************/
+ ******************************************************************************/
 package info.gratour.common.error;
 
 public class ErrorWithCode extends RuntimeException {
@@ -44,17 +45,107 @@ public class ErrorWithCode extends RuntimeException {
     }
 
     public static ErrorWithCode OK = new ErrorWithCode(Errors.OK);
-    public static ErrorWithCode INTERNAL_ERROR = new ErrorWithCode(Errors.INTERNAL_ERROR);
-    public static ErrorWithCode AUTHENTICATION_FAILED = new ErrorWithCode(Errors.AUTHENTICATION_FAILED);
-    public static ErrorWithCode NOT_AUTHENTICATED = new ErrorWithCode(Errors.NOT_AUTHENTICATED);
-    public static ErrorWithCode ACCESS_DENIED = new ErrorWithCode(Errors.ACCESS_DENIED);
-    public static ErrorWithCode NOT_ENOUGH_PRIV = new ErrorWithCode(Errors.NOT_ENOUGH_PRIV);
-    public static ErrorWithCode RECORD_NOT_FOUND = new ErrorWithCode(Errors.RECORD_NOT_FOUND);
-    public static ErrorWithCode EXECUTION_ERROR = new ErrorWithCode(Errors.EXECUTION_ERROR);
-    public static ErrorWithCode ILLEGAL_STATE = new ErrorWithCode(Errors.ILLEGAL_STATE);
-    public static ErrorWithCode INVALID_CONFIG = new ErrorWithCode(Errors.INVALID_CONFIG);
-    public static ErrorWithCode TIMEOUT = new ErrorWithCode(Errors.TIMEOUT);
-    public static ErrorWithCode SERVICE_UNAVAILABLE = new ErrorWithCode(Errors.SERVICE_UNAVAILABLE);
+
+    public static ErrorWithCode internalError() {
+        return new ErrorWithCode(Errors.INTERNAL_ERROR);
+    }
+
+    public static ErrorWithCode internalError(String message) {
+        message = Errors.errorMessage(Errors.INTERNAL_ERROR) + " " + message;
+        return new ErrorWithCode(Errors.INTERNAL_ERROR, message);
+    }
+
+    public static ErrorWithCode internalErrorFmt(String fmt, Object... args) {
+        return internalError(String.format(fmt, args));
+    }
+
+
+    public static ErrorWithCode authenticationFailed() {
+        return new ErrorWithCode(Errors.AUTHENTICATION_FAILED);
+    }
+
+    public static ErrorWithCode notAuthenticated() {
+        return new ErrorWithCode(Errors.NOT_AUTHENTICATED);
+    }
+
+    public static ErrorWithCode accessDenied() {
+        return new ErrorWithCode(Errors.ACCESS_DENIED);
+    }
+
+    public static ErrorWithCode notEnoughPriv() {
+        return new ErrorWithCode(Errors.NOT_ENOUGH_PRIV);
+    }
+
+    public static ErrorWithCode recordNotFound() {
+        return new ErrorWithCode(Errors.RECORD_NOT_FOUND);
+    }
+
+    public static ErrorWithCode recordNotFound(String rec) {
+        return new ErrorWithCode(Errors.RECORD_NOT_FOUND,
+                Errors.errorMessageFormat(Errors.MESSAGE_KEY_RECORD_NOT_FOUND_FMT, rec)
+        );
+    }
+
+    public static ErrorWithCode recordNotFoundFmt(String fmt, Object... args) {
+        return recordNotFound(String.format(fmt, args));
+    }
+
+    public static ErrorWithCode executionError() {
+        return new ErrorWithCode(Errors.EXECUTION_ERROR);
+    }
+
+    public static ErrorWithCode executionError(String detail) {
+        return new ErrorWithCode(
+                Errors.EXECUTION_ERROR,
+                Errors.errorMessageFormat(Errors.MESSAGE_KEY_EXECUTION_ERROR_FMT, detail)
+        );
+    }
+
+    public static ErrorWithCode executionErrorFmt(String fmt, Object... args) {
+        return executionError(String.format(fmt, args));
+    }
+
+    public static ErrorWithCode illegalState() {
+        return new ErrorWithCode(Errors.ILLEGAL_STATE);
+    }
+
+    public static ErrorWithCode illegalState(String explanation) {
+        return new ErrorWithCode(
+                Errors.ILLEGAL_STATE,
+                Errors.errorMessageFormat(Errors.MESSAGE_KEY_ILLEGAL_STATE_FMT, explanation)
+        );
+    }
+
+    public static ErrorWithCode illegalStateFmt(String fmt, Object...args) {
+        return illegalState(String.format(fmt, args));
+    }
+
+    public static ErrorWithCode invalidConfig() {
+        return new ErrorWithCode(Errors.INVALID_CONFIG);
+    }
+
+    public static ErrorWithCode invalidConfig(String configItem) {
+        return new ErrorWithCode(
+                Errors.INVALID_CONFIG,
+                Errors.errorMessageFormat(Errors.MESSAGE_KEY_INVALID_CONFIG_FMT, configItem)
+        );
+    }
+
+    public static ErrorWithCode invalidConfigFmt(String fmt, Object...args) {
+        return invalidParam(String.format(fmt, args));
+    }
+
+    public static ErrorWithCode timeout() {
+        return new ErrorWithCode(Errors.TIMEOUT);
+    }
+
+    public static ErrorWithCode serviceUnavailable() {
+        return new ErrorWithCode(Errors.SERVICE_UNAVAILABLE);
+    }
+
+    public static ErrorWithCode canceled() {
+        return new ErrorWithCode(Errors.CANCELED);
+    }
 
     public static ErrorWithCode invalidParam(String paramName) {
         return new ErrorWithCode(Errors.INVALID_PARAM, Errors.errorMessageFormat(Errors.INVALID_PARAM, paramName));
@@ -66,17 +157,16 @@ public class ErrorWithCode extends RuntimeException {
     }
 
     public static ErrorWithCode invalidValue(String fieldName) {
-        return new ErrorWithCode(Errors.INVALID_VALUE, Errors.errorMessageFormat(Errors.INVALID_VALUE, fieldName));
-    }
-
-    public static ErrorWithCode internalError(String message) {
-        message = Errors.errorMessage(Errors.INTERNAL_ERROR) + " " + message;
-        return new ErrorWithCode(Errors.INTERNAL_ERROR, message);
+        return new ErrorWithCode(
+                Errors.INVALID_VALUE,
+                Errors.errorMessageFormat(Errors.MESSAGE_KEY_INVALID_VALUE_FMT, fieldName)
+        );
     }
 
     public static ErrorWithCode format(int errCode, String formatArg) {
         String message = Errors.errorMessageFormat(errCode, formatArg);
         return new ErrorWithCode(errCode, message);
     }
+
 
 }