+++ doc/change-log.md
@@ -1,6 +1,9 @@
 # v2.0.0
 ## Improvement
+- Maven group changed to `com.lucendar`
 - Move NetUtils from `server-common` to this bundle and translate to java.
+- `com.lucendar.common.utils.DateTimeUtils.BeijingConv` added `millisToOdt`, `millisToOdtBoxed`
+- Introduce `Ref`
 
 # v1.0.0
 - Return type of Reply.ofUpdateResult() changed to `Reply[Void]`